apps/agent-deploy/pyproject.toml
[project]
name = "agent-deploy"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "docker>=7.1.0",
    "flask>=3.0.3",
    "grpcio>=1.71.0",
    "httpx>=0.28.1",
    "kubernetes>=31.0.0",
    "protobuf>=6.32,<7",
]

[build-system]
requires = ["uv_build>=0.8.9,<0.9.0"]
build-backend = "uv_build"

[project.scripts]
serve = "agent_deploy.main:main"


[tool.ruff]
line-length = 120
extend-exclude = ["protos"]

[tool.ruff.format]
docstring-code-format = true

[tool.ruff.lint.isort]
# Use a single line after each import block.
lines-after-imports = 1

[dependency-groups]
dev = ["ruff>=0.12.10"]
